# docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  wechat.JSSDKConfig:
    properties:
      appId:
        type: string
      nonceStr:
        type: string
      signature:
        type: string
      timestamp:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /wechat/jssdk/config:
    get:
      consumes:
      - application/json
      description: 获取用于微信分享等功能的JS-SDK配置信息
      parameters:
      - description: 需要签名的页面URL
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wechat.JSSDKConfig'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取微信JS-SDK配置
      tags:
      - 微信分享
...
swagger: "2.0"